refactor(volar/jsx-directive): code splitting
